single coin for chart api call
